# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: sosbackend
  version: '0.1'
  description: ''
host: 'localhost:3000'
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
paths:
  '/status/{uuid}':
    get:
      summary: ''
      tags:
        - suggested
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'JWE message from an authority '
                pattern: '^[/-A-Za-z0-9+=_.]+$'
              tokenid:
                type: string
                description: 'signaling token id, utilized in the event of abuse '
            required:
              - tokenid
        '404':
          description: Not Found
          schema: {}
      operationId: get-status
      description: 'returns the current status '
    delete:
      summary: ''
      operationId: delete-status
      responses:
        '202':
          description: Accepted
        '401':
          description: Unauthorized
          schema: {}
        '404':
          description: Not Found
          schema: {}
      description: 'clear status, may need a different token authorithy '
      parameters:
        - type: string
          in: header
          name: Authorization
          required: true
          description: deletion token
          pattern: '^[/-A-Za-z0-9+=_.]+$'
    parameters:
      - type: string
        name: uuid
        in: path
        description: 'reference uuid used in encrypted ids '
        maxLength: 0
        minLength: 0
        pattern: '^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$'
        required: true
  /status:
    post:
      summary: ''
      operationId: post-status
      responses:
        '202':
          description: Accepted
        '401':
          description: Unauthorized
          schema:
            type: object
            properties: {}
      parameters:
        - type: string
          in: header
          name: Authorization
          required: true
          description: 'JWT one time validity! '
          pattern: '^[/-A-Za-z0-9+=_.]+$'
        - in: body
          name: body
          schema:
            type: object
            properties:
              refids:
                type: array
                description: JWE ref-IDs the device saw
                minItems: 1
                items:
                  type: string
                  pattern: '^[/-A-Za-z0-9+=_.]+$'
              message:
                type: string
                pattern: '^[/-A-Za-z0-9+=_.]+$'
                description: JWE with a nice message from some authorithy
          description: ''
      description: "push your seen reference IDs and the \"message\" \ncascade with other systems -> tokenid based idempotency "
      tags:
        - required
  /.well-known/jwks.json:
    get:
      summary: ''
      tags:
        - suggested
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/jsonwebkeyset'
      operationId: get-publickeys
      description: 'returns all well known public keys '
    parameters: []
  /token:
    get:
      summary: ''
      tags:
        - authority
      responses:
        '200':
          description: OK
          schema:
            type: string
            description: JWT-token ;)
            pattern: '^[/-A-Za-z0-9+=_.]+$'
        '500':
          description: Internal Server Error
          schema: {}
      operationId: get-token
      description: 'returns a simple token, spice it up with the surveilance and authority you like.'
definitions:
  jsonwebkeyset:
    type: object
    properties:
      keys:
        description: list of keys
        type: array
        items:
          type: object
          description: json web key
          properties:
            'n':
              type: string
              description: modulus
            e:
              type: string
              description: exponent
            kty:
              type: string
              description: ktype
            alg:
              type: string
              description: alorithm
            kid:
              type: string
              description: keyid
            k:
              type: string
              description: key
          required:
            - kty
            - kid